jQuery and other corrections

+++ src/Controller/ImageController.php
@@ -21,24 +21,25 @@ class ImageController extends AbstractController
      */
     public function delete(Image $image, Request $request, TrickRepository $trickRepository)
     {
-        dump($request->request->all(), $image, $this->isCsrfTokenValid('delete'.$image->getId(), $request->request->get('token')));
         $trick = $trickRepository->find($request->request->get('trick'));
-        $changed = false;
 
+        if (!$trick || !$image) {
+            return new JsonResponse(['error' => true]);
+        }
+
+        $changed = false;
 
         if ($this->isCsrfTokenValid('delete'.$image->getId(), $request->request->get('token'))) {
             if ($trick->getMainImage() === $image) {
                 $images = $trick->getImages();
                 $images->removeElement($image);
                 $images->isEmpty() ? $trick->setMainImage(null) : $trick->setMainImage($images->first());
-                $changed = $trick->getMainImage() ? $trick->getMainImage()->getImageName() : 'empty';
+                $changed = $trick->getMainImage() ? $trick->getMainImage()->getImageName() : 'placehold.jpg';
             }
             $em = $this->getDoctrine()->getManager();
             $em->remove($image);
             $em->flush();
         }
-
-        // @todo send error
 
         return new JsonResponse(['changed' => $changed]);
     }
@@ -48,19 +49,16 @@ class ImageController extends AbstractController
      */
     public function replaceMainImg(Image $image, Request $request, TrickRepository $trickRepository)
     {
-        $result = 'error';
         $trick = $trickRepository->find($request->request->get('trick'));
 
         if ($trick && $this->isCsrfTokenValid('delete'.$image->getId(), $request->request->get('token'))) {
-
             $trick->setMainImage($image);
             $em = $this->getDoctrine()->getManager();
             $em->persist($trick);
             $em->flush();
-            $result = 'success';
+            return new JsonResponse(['error' => false]);
         }
-        // @todo if not trick or image send error
-        return new JsonResponse(['result' => $result]);
+        return new JsonResponse(['error' => true]);
     }
 
     /**
@@ -69,13 +67,16 @@ class ImageController extends AbstractController
     public function create(Request $request, ImageRepository $imageRepository, TrickRepository $trickRepository, EntityManagerInterface $entityManager)
     {
         $trick = $trickRepository->find($request->request->get('trick'));
-        // @todo if not trick
+
+        if (!$trick) {
+            return new JsonResponse(['error' => true]);
+        }
+
         $files = $request->files->all();
         $images = new ArrayCollection();
         $changed = false;
 
         foreach ($files as $file) {
-            // @todo check if all images are uploaded or send error
             $image = new Image();
             $image->setImageFile($file)->setTrick($trick);
             if($trick->getMainImage() === null) {
@@ -85,9 +86,13 @@ class ImageController extends AbstractController
             $images->add($image);
         }
 
-        $entityManager->flush();
+        try {
+            $entityManager->flush();
+            $view = $this->renderView('parts/list-images.html.twig', ['images' => $images]);
+            return new JsonResponse(['view' => $view, 'changed' => $changed]);
+        } catch(\Exception $e) {
+            return new JsonResponse(['error' => true]);
+        }
 
-        $view = $this->renderView('parts/list-images.html.twig', ['images' => $images]);
-        return new JsonResponse(['view' => $view, 'changed' => $changed]);
     }
 }
